requirements.txt
fastapi
uvicorn
jinja2
psycopg2-binary



